% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.download.R
\name{add.download}
\alias{add.download}
\title{Add a download command to the bot.}
\usage{
add.download(
  x = Settings,
  command = "download",
  description = "download the database.csv file",
  db.name = "database"
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{command}{Command to type on telegram.}

\item{description}{A description for the command.}

\item{db.name}{Name of the database.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Add a download command to the bot.
}
\examples{
Settings <- add.start.f(Settings, command="download", description="download the database.csv file", db.name ="database")

}
