% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.legend.R
\name{add.legend}
\alias{add.legend}
\title{Add a legend command to the bot.}
\usage{
add.legend(
  x = Settings,
  command = "legend",
  description = "get the list of tasks",
  mss = "The tasks descriptions are:",
  team = "none"
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{command}{Command to type on telegram.}

\item{description}{A description for the command.}

\item{mss}{Message shown on the chat.}

\item{team}{Filter the team from the tasks shown.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Add a legend command to the bot.
}
\examples{
Settings <- add.legend(Settings, command="legend", description="get the list of tasks", mss= "The tasks descriptions are:", team="none")

}
