% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.map.R
\name{add.map}
\alias{add.map}
\title{Add a map command to the bot.}
\usage{
add.map(
  x = Settings,
  command = "tasksmap",
  db.name = "database",
  description = "map of tasks done and to do today",
  type = "User",
  team = NA,
  today = 0,
  title = "Tasks today",
  completed = "done",
  caption = "Tasks to do and done today",
  mss.empty = "No tasks for today"
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{command}{Command to type on telegram..}

\item{db.name}{Name of the database. Should be previously integred to Settings with the @backref set.database function.}

\item{description}{A description for the command.}

\item{type}{Define the type of map. Could be one of the followings: "Tasks", "Status", "User", or the name from a column of the database.}

\item{team}{Filter map content by a team. Culd be NA or a team from the Settings file.}

\item{today}{Number of days added or substracted to current date.}

\item{title}{Title for the map, if NA, there is no title."today" or "tomorrow" will be substituted by real date.}

\item{completed}{Could be "done", to show just done stuff, "to-do" or "all".}

\item{caption}{Caption for the map.}

\item{mss.empty}{Message for when there is no content to show on the map.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Add a map command to the bot.
}
\examples{
Settings <- add.map(Settings, command="donetoday", db.name ="database", description="map of tasks done today", type="User", team=NA, today= 0, title=NA, completed="done", caption="Tasks done today", mss.empty="No tasks done today")

Settings <- add.map(Settings, command="tomorrowmap", db.name ="database", description="map of tasks to do tomorrow", type="Tasks", team=NA, today= 1, title="Tasks today", completed="to-do", caption="Tasks to do tomorrow", mss.empty="No tasks for tomorrow")

Settings <- add.map(Settings, command="statusmap", db.name ="database", description="map of nestboxes status yesterday", type="Status", team=NA, today= -1, title=NA, completed="all", caption="Status of the nestboxes yesterday", mss.empty="No status registered")

}
