% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.messages.R
\name{add.messages}
\alias{add.messages}
\title{Add a messages handler to the bot.}
\usage{
add.messages(
  x = Settings,
  pic.folder = "./Images/",
  mss.saved.pic = "Picture saved!",
  mss.caption.pic = "Please send the picture again with a caption."
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{image.folder}{Folder to save pictures.}

\item{mss.saved}{Message shown when picture saved.}

\item{mss.caption}{Requests a caption to title the picture with.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Add a messages handler to the bot.
}
\examples{
Settings <- add.sampling(Settings, pic.folder="./Images/", mss.saved.pic="Picture saved!", mss.caption.pic="Please send the picture again with a caption.")

}
