% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.newuser.R
\name{add.newuser}
\alias{add.newuser}
\title{Add a newuser registration command to the bot.}
\usage{
add.newuser(
  x = Settings,
  command = "newuser",
  description = "Command to request for access.",
  mss.requested = "You have already requested for access.",
  mss.allowed = "You are already allowed to start using the bot!",
  mss.sent = "... \\n Your request has been sent!"
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{command}{Command to type on telegram.}

\item{description}{A description for the command.}

\item{mss.requested}{"You have already requested for access."}

\item{mss.allowed}{"You are already allowed to start using the bot!"}

\item{mss.sent}{"Your request has been sent!"}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Add a newuser registration command to the bot.
}
\examples{
Settings <- add.legend(x=Settings, command="newuser", description= "Command to request for access.", mss.requested="You have already requested for access.", mss.allowed="You are already allowed to start using the bot!", mss.sent="Your request has been sent!")

}
