% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.sampling.R
\name{add.sampling}
\alias{add.sampling}
\title{Add a sampling command to the bot.}
\usage{
add.sampling(
  x = Settings,
  command = "nest",
  db.name = "database",
  description = "add info to each nest for today",
  sep = "-",
  today = 0,
  type = "Info",
  mss.filled = " info filled.",
  mss.tasks = "The task for today was: ",
  mss.rewrite = "Please use command /rewrite instead of /nest to rewrite wrong info.",
  mss.previous = " was previously filled as:"
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{command}{Command to type on telegram. .}

\item{db.name}{Name of the database. Should be previously integred to Settings with the @backref set.database function.}

\item{description}{A description for the command.}

\item{sep}{Define the separator to write between the nestbox number and the info.}

\item{today}{Number of days added or substracted to current date.}

\item{type}{Type of data to imput. Should be one of 'Tasks', 'Status' or 'Info'.}

\item{mss.filled}{Message when info filled.}

\item{mss.tasks}{Message to show the task.}

\item{mss.rewrite}{Message info is already imputed.}

\item{mss.previous}{Message to show previous info.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Add a sampling command to the bot.
}
\examples{
Settings <- add.sampling(Settings, command="nest", db.name ="database", description="add info to each nest for today", sep="-", today= 0, type="Info", mss.filled=" info filled.", mss.tasks="The task for today was: ", mss.rewrite="Please use command /rewrite instead of /nest to rewrite wrong info.", mss.previous=" was previously filled as:")

Settings <- add.sampling(Settings, command="yesterday-nest", db.name ="database", description="add info to each nest for yesterday", sep=".", today= -1, type="Info", mss.filled=" info filled.", mss.tasks="The task for yesterday was: ", mss.rewrite="Info for yesterday already imputed", mss.previous=" was previously filled as:")
}
