% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.start.R
\name{add.start}
\alias{add.start}
\title{Add a start command to the bot. Adding a start command is recommended.}
\usage{
add.start(
  x = Settings,
  mss.welcome = "Welcome to the starlings data collection bot. 😊",
  mss.commands = "There are some commands you should learn:"
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{mss.welcome}{First message shown.}

\item{mss.commands}{Second message shown.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Add a start command to the bot. Adding a start command is recommended.
}
\examples{
Settings <- add.start(Settings, mss.welcome= "Welcome to the starlings data collection bot.", mss.commands="There are some commands you should learn:")

}
