% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.view.R
\name{add.view}
\alias{add.view}
\title{Add a view command to the bot.}
\usage{
add.view(
  x = Settings,
  command = "view",
  db.name = "database",
  description = "view current days of a nestbox",
  show = c("Tasks", "Status", "Info"),
  firstday = -5,
  lastday = 1,
  caption = "Last days of nestbox ",
  mss.empty = "No data on previous week for this nest."
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{command}{Command to type on telegram. .}

\item{db.name}{Name of the database. Should be previously integred to Settings with the @backref set.database function.}

\item{description}{A description for the command.}

\item{show}{what to show on the table, could be 'c("Tasks","Status","Info")'.}

\item{firstday}{Number of days added or substracted to current date in which the table should start.}

\item{lastday}{Number of days added or substracted to current date in which the table should end.}

\item{caption}{Caption for the image.}

\item{mss.empty}{Message when there is no data.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Add a view command to the bot.
}
\examples{
Settings <- add.view(Settings, command="viewweek", db.name ="database", description="view last week status of a nestbox", show="Status", firstday= -7, lastday= 0, caption="Last week status of nestbox ", mss.empty="No data on previous week for this nest.")

}
