% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.to.database.R
\name{calendar.to.database}
\alias{calendar.to.database}
\title{Function to refresh the database with the calendar excel file.}
\usage{
calendar.to.database(x = Settings, db.name = "database")
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{db.name}{Name of the database. Should be previously integred to Settings with the @backref set.database function.}

\item{command}{A number.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Function to refresh the database with the calendar excel file.
}
\examples{
Settings <- calendar.to.database(Settings, db.name = "database")

}
