% Generated by roxygen2: do not edit by hand
% Please edit documentation in ./calendar.to.database
\name{database.to.calendar}
\alias{database.to.calendar}
\title{Turns the database file into a calendar format excel file. It makes editing easier. Make sure you sve changes on the database with the 'calendar.to.database' function.}
\usage{
database.to.calendar(x = Settings, db.name = "database")
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{db.name}{Name of the database. Should be previously integred to Settings with the @backref set.database function.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Turns the database file into a calendar format excel file. It makes editing easier. Make sure you sve changes on the database with the 'calendar.to.database' function.
}
\examples{
Settings <- database.to.calendar(Settings, db.name = "database")

}
