% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.color.df.R
\name{get.color.df}
\alias{get.color.df}
\title{Get the colors from a dataframe.}
\usage{
get.color.df(dir, sheet.)
}
\arguments{
\item{dir}{Directory of the excel file.}

\item{sheet.}{Sheet name.}
}
\value{
A dataframe filled with colors.
}
\description{
Get the colors from a dataframe.
}
