% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.Bot.R
\name{run.Bot}
\alias{run.Bot}
\title{Function to run the Bot. Mantain the computer on while the bot should be runing.}
\usage{
run.Bot(x = Settings)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}
}
\description{
Function to run the Bot. Mantain the computer on while the bot should be runing.
}
\examples{
run.Bot(x=Settings)

}
