% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.database.R
\name{set.database}
\alias{set.database}
\title{Crates a database file with a row for each date and nestbox.}
\usage{
set.database(
  x = Settings,
  db.name = "database",
  firstday = "2024-10-28",
  lastday = "2024-11-06"
)
}
\arguments{
\item{x}{A 'BirdBotSettings' object.}

\item{db.name}{Name of the database.}

\item{firstday}{Date in which the database will start.}

\item{lastday}{Date in which the database will end}

\item{command}{A number.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Crates a database file with a row for each date and nestbox.
}
\examples{
Settings <- set.database(Settings, db.name = "database", firstday = "2024-10-28", lastday = "2024-11-06")

}
