% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.settings.R
\name{set.settings}
\alias{set.settings}
\title{Sets Settings for the Bot. It will create a Settings file or refresh an existing one.}
\usage{
set.settings(
  dir = "./data/Settings.xlsx",
  date.format = "\%d/\%m/\%Y",
  map.file = "./data/StudyArea.kml",
  mss.deny = "You are not allowed to use this bot. 🚫",
  mss.request = "Please request access by sending a \\"/newuser\\" message.",
  mss.wrong.nest = "Please, enter a valid nestbox number.",
  mss.no.nest = "Please, enter a nestbox number.",
  mss.unknown = "I did not understand. 😰 Please, repeat."
)
}
\arguments{
\item{dir}{A 'BirdBotSettings' object.}

\item{date.format}{The date format used on excel files.}

\item{map.file}{Directory for the '.klm' map file. Could be NA if you do not have a file.}

\item{mss.deny}{Message to deny access to the bot.}

\item{mss.request}{Message to request for access to the bot.}

\item{mss.wrong.nest}{Message if incorrect nestbox number is typed.}

\item{mss.no.nest}{Message if no nestbox number is typed.}

\item{mss.unknown}{Ununderstood message.}
}
\value{
A 'BirdBotSettings' object.
}
\description{
Sets Settings for the Bot. It will create a Settings file or refresh an existing one.
}
\examples{
Settings <- set.settings(dir ="./data/Settings.xlsx", date.format="\%d/\%m/\%Y", map.file="./data/StudyArea .kml")

}
